Add PALSAR example downloaded into the TESTDATA_DIR
To test the PALSAR functionality we are downloading an example file from
ftp://ftp.eorc.jaxa.jp/
The download is handled in the download_testdata.sh script.
This script is invoked in the before_script phase of travis.
The test_drivers file uses the TESTDATA_DIR environement variable to
determine where the testdata is located.